+++ wgpu-core-remote/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "wgpu-core-remote"
+version.workspace = true
+edition.workspace = true
+description = "Remoting version of wgpu-core"
+homepage.workspace = true
+repository.workspace = true
+keywords.workspace = true
+license.workspace = true
+
+# Override the workspace's `rust-version` key. `wgpu-core-remote` and its dependencies
+# have a less strict MSRV, to allow firefox more leeway in updating their Rust toolchain.
+#
+# See the repo README for more information on MSRV policy.
+rust-version = "1.87"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+targets = [
+    "x86_64-unknown-linux-gnu",
+    "x86_64-apple-darwin",
+    "x86_64-pc-windows-msvc",
+    "wasm32-unknown-unknown",
+]
+
+[features]
+## Enables serialization via `serde` on common wgpu types.
+serde = ["dep:serde", "wgpu-types/serde", "wgpu-core/serde"]
+
+[dependencies]
+wgpu-core = { workspace = true, default-features = false }
+wgpu-hal = { workspace = true, default-features = false }
+wgpu-types.workspace = true
+parking_lot.workspace = true
+raw-window-handle.workspace = true
+serde = { workspace = true, optional = true }
