+++ wgpu-core/src/command/transfer.rs
@@ -14,8 +14,6 @@ use crate::{
         clear_texture, encoder::EncodingState, ArcCommand, CommandEncoderError, EncoderStateError,
     },
     device::MissingDownlevelFlags,
-    global::Global,
-    id::{BufferId, CommandEncoderId, TextureId},
     init_tracker::{
         has_copy_partial_init_tracker_coverage, MemoryInitKind, TextureInitRange,
         TextureInitTrackerAction,
@@ -28,7 +26,6 @@ use crate::{
 
 use super::ClearError;
 
-type TexelCopyBufferInfo = wgt::TexelCopyBufferInfo<BufferId>;
 type TexelCopyTextureInfo = wgt::TexelCopyTextureInfo<Arc<Texture>>;
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -964,92 +961,6 @@ impl super::CommandEncoder {
                 size: *copy_size,
             })
         })
-    }
-}
-
-impl Global {
-    pub fn command_encoder_copy_buffer_to_buffer(
-        &self,
-        command_encoder_id: CommandEncoderId,
-        source: BufferId,
-        source_offset: BufferAddress,
-        destination: BufferId,
-        destination_offset: BufferAddress,
-        size: Option<BufferAddress>,
-    ) -> Result<(), EncoderStateError> {
-        let hub = &self.hub;
-
-        let cmd_enc = hub.command_encoders.get(command_encoder_id);
-        let source = self.resolve_buffer_id(source);
-        let destination = self.resolve_buffer_id(destination);
-        cmd_enc.copy_buffer_to_buffer(source, source_offset, destination, destination_offset, size)
-    }
-
-    pub fn command_encoder_copy_buffer_to_texture(
-        &self,
-        command_encoder_id: CommandEncoderId,
-        source: &TexelCopyBufferInfo,
-        destination: &wgt::TexelCopyTextureInfo<TextureId>,
-        copy_size: &Extent3d,
-    ) -> Result<(), EncoderStateError> {
-        let cmd_enc = self.hub.command_encoders.get(command_encoder_id);
-        let source = wgt::TexelCopyBufferInfo {
-            buffer: self.resolve_buffer_id(source.buffer),
-            layout: source.layout,
-        };
-        let destination = wgt::TexelCopyTextureInfo {
-            texture: self.resolve_texture_id(destination.texture),
-            mip_level: destination.mip_level,
-            origin: destination.origin,
-            aspect: destination.aspect,
-        };
-        cmd_enc.copy_buffer_to_texture(&source, &destination, copy_size)
-    }
-
-    pub fn command_encoder_copy_texture_to_buffer(
-        &self,
-        command_encoder_id: CommandEncoderId,
-        source: &wgt::TexelCopyTextureInfo<TextureId>,
-        destination: &TexelCopyBufferInfo,
-        copy_size: &Extent3d,
-    ) -> Result<(), EncoderStateError> {
-        let cmd_enc = self.hub.command_encoders.get(command_encoder_id);
-
-        let source = wgt::TexelCopyTextureInfo {
-            texture: self.resolve_texture_id(source.texture),
-            mip_level: source.mip_level,
-            origin: source.origin,
-            aspect: source.aspect,
-        };
-        let destination = wgt::TexelCopyBufferInfo {
-            buffer: self.resolve_buffer_id(destination.buffer),
-            layout: destination.layout,
-        };
-        cmd_enc.copy_texture_to_buffer(&source, &destination, copy_size)
-    }
-
-    pub fn command_encoder_copy_texture_to_texture(
-        &self,
-        command_encoder_id: CommandEncoderId,
-        source: &wgt::TexelCopyTextureInfo<TextureId>,
-        destination: &wgt::TexelCopyTextureInfo<TextureId>,
-        copy_size: &Extent3d,
-    ) -> Result<(), EncoderStateError> {
-        let cmd_enc = self.hub.command_encoders.get(command_encoder_id);
-
-        let source = wgt::TexelCopyTextureInfo {
-            texture: self.resolve_texture_id(source.texture),
-            mip_level: source.mip_level,
-            origin: source.origin,
-            aspect: source.aspect,
-        };
-        let destination = wgt::TexelCopyTextureInfo {
-            texture: self.resolve_texture_id(destination.texture),
-            mip_level: destination.mip_level,
-            origin: destination.origin,
-            aspect: destination.aspect,
-        };
-        cmd_enc.copy_texture_to_texture(&source, &destination, copy_size)
     }
 }
