+++ wgpu-core/src/lib.rs
@@ -75,10 +75,7 @@ pub mod command;
 mod conv;
 pub mod device;
 pub mod error;
-pub mod global;
-pub mod hub;
 pub mod id;
-pub mod identity;
 mod indirect_validation;
 mod init_tracker;
 pub mod instance;
@@ -89,7 +86,6 @@ mod pipeline_cache;
 mod pool;
 pub mod present;
 pub mod ray_tracing;
-pub mod registry;
 pub mod resource;
 mod snatch;
 pub mod storage;
@@ -120,13 +116,10 @@ pub(crate) use nt::{FastHashMap, FastHashSet, FastIndexMap};
 /// These are the values stored in `Device::fence`.
 pub type SubmissionIndex = hal::FenceValue;
 
-type Index = u32;
-type Epoch = u32;
-
 pub type RawString = *const core::ffi::c_char;
 pub type Label<'a> = Option<Cow<'a, str>>;
 
-trait LabelHelpers<'a> {
+pub trait LabelHelpers<'a> {
     fn to_hal(&'a self, flags: wgt::InstanceFlags) -> Option<&'a str>;
     fn to_string(&self) -> String;
 }
