+++ wgpu-core/src/present.rs
@@ -17,8 +17,7 @@ use crate::device::trace::{Action, IntoTrace};
 use crate::{
     conv,
     device::{queue::Queue, Device, DeviceError, MissingDownlevelFlags, WaitIdleError},
-    global::Global,
-    hal_label, id,
+    hal_label,
     instance::Surface,
     resource::{self, Labeled},
 };
@@ -479,45 +478,5 @@ impl Surface {
             .ok_or(SurfaceError::NothingToPresent)?;
 
         Ok(())
-    }
-}
-
-impl Global {
-    pub fn surface_get_current_texture(
-        &self,
-        surface_id: id::SurfaceId,
-        texture_id_in: Option<id::TextureId>,
-    ) -> Result<SurfaceOutput<id::TextureId>, SurfaceError> {
-        let surface = self.surfaces.get(surface_id);
-
-        let fid = self.hub.textures.prepare(texture_id_in);
-
-        let output = surface.get_current_texture()?;
-
-        let status = output.status;
-        let texture_id = output.texture.map(|texture| fid.assign(texture));
-
-        Ok(SurfaceOutput {
-            status,
-            texture: texture_id,
-        })
-    }
-
-    pub fn surface_present(&self, surface_id: id::SurfaceId) -> Result<Status, SurfaceError> {
-        let surface = self.surfaces.get(surface_id);
-
-        surface.present()
-    }
-
-    pub fn surface_texture_discard(&self, surface_id: id::SurfaceId) -> Result<(), SurfaceError> {
-        let surface = self.surfaces.get(surface_id);
-
-        surface.discard()
-    }
-
-    pub fn surface_texture_release(&self, surface_id: id::SurfaceId) -> Result<(), SurfaceError> {
-        let surface = self.surfaces.get(surface_id);
-
-        surface.release()
     }
 }
